baseline for error handling changes
