+++ src/lexer/mod.rs
@@ -3,5 +3,5 @@ pub mod lexer;
 mod tests;
 
 pub mod prelude {
-    pub use crate::lexer::{lexer::{Lexer, LexerIterator}, token::{Token, TokenKind, TokenSpan}};
+    pub use crate::lexer::{lexer::{Lexer, LexerIterator}, token::{Token, TokenKind}};
 }
