+++ src/parser/mod.rs
@@ -1,5 +1,7 @@
 mod ast;
 mod parser;
+mod combinator;
+mod error;
 mod ptr;
 mod tests;
 
@@ -10,6 +12,6 @@ use crate::lexer::token::Token;
 use parser::*;
 
 // Parser entry point.
-pub fn parse<'a>(input: &'a [Token]) -> ParseResult<'a, (ast::Module, Vec<parser::ParserError>)> {
+pub fn parse<'a>(input: &'a [Token]) -> ParseResult<'a, ast::Module> {
     parse_module().parse(input)
 }
